Reorg and rename fixed buffers API
Rename BufRegistry to FixedBufRegistry.
Move most of the internal buffer infrastructure as submodules
under buf::fixed. Rename driver::buffers to driver::register_buffers.
