+++ src/buf/fixed/handle.rs
@@ -1,0 +1,74 @@
+use super::FixedBuffers;
+use crate::buf::{IoBuf, IoBufMut};
+
+use libc::iovec;
+use std::cell::RefCell;
+use std::mem::ManuallyDrop;
+use std::rc::Rc;
+
+/// A unique handle to a memory buffer that can be pre-registered with
+/// the kernel for `io-uring` operations.
+///
+/// `FixedBuf` handles can be obtained from a [`BufRegistry`] collection.
+/// For each buffer, only a single `FixedBuf` handle can be either used by the
+/// application code or owned by an I/O operation at any given time,
+/// thus avoiding data races between `io-uring` operations in flight and
+/// the application accessing buffer data.
+pub struct FixedBuf {
+    registry: Rc<RefCell<FixedBuffers>>,
+    buf: ManuallyDrop<Vec<u8>>,
+    index: u16,
+}
+
+impl Drop for FixedBuf {
+    fn drop(&mut self) {
+        let mut registry = self.registry.borrow_mut();
+        registry.check_in(self.index as usize, self.buf.len());
+    }
+}
+
+impl FixedBuf {
+    pub(super) unsafe fn new(
+        registry: Rc<RefCell<FixedBuffers>>,
+        iovec: iovec,
+        init_len: usize,
+        index: u16,
+    ) -> Self {
+        let buf = Vec::from_raw_parts(iovec.iov_base as _, init_len, iovec.iov_len);
+        FixedBuf {
+            registry,
+            buf: ManuallyDrop::new(buf),
+            index,
+        }
+    }
+
+    pub(crate) fn buf_index(&self) -> u16 {
+        self.index
+    }
+}
+
+unsafe impl IoBuf for FixedBuf {
+    fn stable_ptr(&self) -> *const u8 {
+        self.buf.as_ptr()
+    }
+
+    fn bytes_init(&self) -> usize {
+        self.buf.len()
+    }
+
+    fn bytes_total(&self) -> usize {
+        self.buf.capacity()
+    }
+}
+
+unsafe impl IoBufMut for FixedBuf {
+    fn stable_mut_ptr(&mut self) -> *mut u8 {
+        self.buf.as_mut_ptr()
+    }
+
+    unsafe fn set_init(&mut self, pos: usize) {
+        if self.buf.len() < pos {
+            self.buf.set_len(pos)
+        }
+    }
+}
